feat: add windows specific instructions
