+++ internal/commons/anvil.go
@@ -1,0 +1,10 @@
+//go:build !windows
+
+package commons
+
+import "context"
+
+// Prerequisites implements HandleRelease.
+func (a *AnvilRelease) Prerequisites(ctx context.Context) error {
+	return nil
+}
